requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
langchain-community
langchain-huggingface
langchain_mongodb
faiss-cpu
groq
requests
python-decouple
scikit-learn
sentence-transformers
python-telegram-bot
pymongo
mongoengine
starlette-admin
passlib[bcrypt]
python-jose
pyjwt
itsdangerous
